# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.GetUserInfoResponseData:
    properties:
      id:
        example: "1234567890"
        type: string
      needReset:
        example: false
        type: boolean
      role:
        example: admin
        type: string
      username:
        example: user123
        type: string
    type: object
  dto.LoginResponseData:
    properties:
      token:
        description: JWT token
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 用户信息
    type: object
  dto.ResetPasswordResponseData:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.User'
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  dto.UserCreateRequest:
    properties:
      password:
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - password
    - role
    - username
    type: object
  dto.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UserPasswordResetRequest:
    properties:
      newPassword:
        minLength: 6
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  dto.UserUpdateRequest:
    properties:
      needReset:
        type: boolean
      password:
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - auditor
        - configurator
        - user
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    type: object
  model.APIResponse:
    description: API响应的标准格式
    properties:
      code:
        description: HTTP状态码
        example: 200
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息(当success=false时)
        example: 参数错误
        type: string
      message:
        description: 响应消息
        example: 操作成功
        type: string
      requestId:
        description: 请求ID，用于跟踪
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        description: 是否成功
        example: true
        type: boolean
      timestamp:
        description: 响应时间戳
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.ErrResponse:
    description: 错误的API响应标准格式
    properties:
      code:
        example: 400
        type: integer
      error:
        example: 参数错误
        type: string
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.ErrResponseDontShowError:
    description: 错误的API响应标准格式,不展示 error
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.SuccessResponse:
    description: 成功的API响应标准格式
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.SuccessResponseNoData:
    description: 成功的API响应标准格式,没有 data
    properties:
      code:
        example: 200
        type: integer
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastLogin:
        type: string
      needReset:
        description: 是否需要重置密码
        type: boolean
      permissions:
        description: 额外权限
        items:
          type: string
        type: array
      role:
        description: 用户角色
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
host: localhost:2333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 简单的 Web 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple-WAF API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取JWT令牌
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponseData'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      summary: 用户登录
      tags:
      - 认证
  /auth/me:
    get:
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户信息成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetUserInfoResponseData'
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: 用户重置自己的密码
      parameters:
      - description: 密码重置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码重置成功
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "400":
          description: 请求参数错误或原密码错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 重置密码
      tags:
      - 认证
  /users:
    get:
      description: 获取系统中所有用户的列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户列表成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取所有用户
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 管理员创建新用户
      parameters:
      - description: 用户创建信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户创建成功
          schema:
            $ref: '#/definitions/dto.ResetPasswordResponseData'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "409":
          description: 用户名已存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 创建新用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      description: 管理员删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户删除成功
          schema:
            $ref: '#/definitions/model.APIResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 管理员更新指定用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户更新成功
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.APIResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"